+++ lib/advanced_api_client.dart
@@ -5,4 +5,5 @@ export 'src/core/refresh_config.dart';
 export 'src/interceptors/auth_interceptor.dart';
 export 'src/interceptors/retry_interceptor.dart';
 export 'src/advanced_api_client_impl.dart';
+export 'src/utils/api_logger.dart';
 export 'src/storage/shared_pref_token_storage.dart';