+++ lib/src/core/refresh_config.dart
@@ -1,12 +1,16 @@
+import 'dart:async';
+
 import 'package:dio/dio.dart';
 
 typedef RefreshTokenParser = String? Function(dynamic responseData);
+typedef RefreshBodyBuilder = FutureOr<Map<String, dynamic>> Function();
 
 class RefreshConfig {
   final String path;
   final String method;
   final Map<String, dynamic>? query;
-  final dynamic body;
+  final Map<String, dynamic>? body; // static body
+  final RefreshBodyBuilder? bodyBuilder; // dynamic body
   final Map<String, dynamic>? headers;
   final RefreshTokenParser tokenParser;
 
@@ -16,6 +20,7 @@ class RefreshConfig {
     required this.tokenParser,
     this.query,
     this.body,
+    this.bodyBuilder,
     this.headers,
   });